requirements.txt
django==4.1.2
django-cleanup==6.0.0
django-extensions==3.2.0
django-crispy-forms==1.14.0
